move wrapper code out of the way
